Code cleanup and housekeeping
